refactor: func_expr をモジュール化
