+++ crates/uroborosql-fmt/src/new_visitor/pg_expr/c_expr.rs
@@ -1,4 +1,4 @@
-mod function;
+mod func_expr;
 
 use postgresql_cst_parser::{syntax_kind::SyntaxKind, tree_sitter::TreeCursor};
